perf: 优化初始化脚本

+++ init.sh
@@ -1,17 +1,28 @@
 #!/bin/bash
 #
 
+# helm-charts mount
+# https://github.com/jumpserver/helm-charts/blob/main/charts/jumpserver/templates/web/deployment-nginx.yaml#L60
+if [ -f "/etc/nginx/sites-enabled/jms.conf" ]; then
+  cp -rf /etc/nginx/sites-enabled/jms.conf /etc/nginx/conf.d/default.conf
+fi
+
 if [ -n "${CLIENT_MAX_BODY_SIZE}" ]; then
   sed -i "s@client_max_body_size .*;@client_max_body_size ${CLIENT_MAX_BODY_SIZE};@g" /etc/nginx/conf.d/default.conf
 fi
-if [ ! "${HTTPS_PORT}" ] && [ "${HTTP_PORT}" != "80" ]; then
-  sed -i "s@listen 80;@listen ${HTTP_PORT};@g" /etc/nginx/conf.d/default.conf
-fi
-if [ ! "${HTTPS_PORT}" ] && [ "${USE_IPV6}" == "1" ]; then
-  sed -i "s@# listen \[::\]:80;@listen \[::\]:${HTTP_PORT};@g" /etc/nginx/conf.d/default.conf
+
+if [ -n "${HTTP_PORT}" ]; then
+  if [ ! "${HTTPS_PORT}" ] && [ "${HTTP_PORT}" != "80" ]; then
+    sed -i "s@listen 80;@listen ${HTTP_PORT};@g" /etc/nginx/conf.d/default.conf
+  fi
+  if [ ! "${HTTPS_PORT}" ] && [ "${USE_IPV6}" == "1" ]; then
+    sed -i "s@# listen \[::\]:80;@listen \[::\]:${HTTP_PORT};@g" /etc/nginx/conf.d/default.conf
+  fi
 fi
 
-if [ -f "/etc/nginx/sites-enabled/jumpserver.conf" ]; then
+# Installer mount
+# https://github.com/jumpserver/installer/blob/dev/compose/docker-compose-lb.yml#L14
+if [ -n "${HTTPS_PORT}" ] && [ -f "/etc/nginx/sites-enabled/jumpserver.conf" ]; then
   cp -rf /etc/nginx/sites-enabled/jumpserver.conf /etc/nginx/conf.d/jumpserver.conf
   sed -i "s@listen 80;@listen 8080;@g" /etc/nginx/conf.d/default.conf
   sed -i "s@server web:80;@server web:8080;@g" /etc/nginx/conf.d/jumpserver.conf
